renames journalistdashboard to dashboard refactors article js
