+++ src/state/reducers/rootReducer.js
@@ -1,27 +1,27 @@
 const rootReducer = (state, action) => {
-  switch (action.type) {
-    case 'SUCCESS_MESSAGE':
-      return {
-        ...state,
-        message: action.payload,
-        successOpen: true,
-      };
-    case 'CLOSE_MESSAGE':
-      return {
-        ...state,
-        successOpen: false,
-        errorOpen: false,
-      };
-    case 'ERROR_MESSAGE':
-      return {
-        ...state,
-        message: action.payload,
-        errorOpen: true,
-      };
+	switch (action.type) {
+		case 'SUCCESS_MESSAGE':
+			return {
+				...state,
+				message: action.payload,
+				successOpen: true,
+			};
+		case 'CLOSE_MESSAGE':
+			return {
+				...state,
+				successOpen: false,
+				errorOpen: false,
+			};
+		case 'ERROR_MESSAGE':
+			return {
+				...state,
+				message: action.payload,
+				errorOpen: true,
+			};
 
-    default:
-      return state;
-  }
+		default:
+			return state;
+	}
 };
 
 export default rootReducer;
